simplified the examples for testing and changed the config files
