--- test/interaction/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "geojson",
-  "version": "0.0.0",
-  "license": "MIT",
-  "scripts": {
-    "start-local": "webpack-dev-server --env.local --progress --hot --open",
-    "start": "webpack-dev-server --progress --hot --open"
-  },
-  "dependencies": {
-    "d3-request": "^1.0.5",
-    "is-little-endian": "0.0.0",
-    "d3-scale": "^1.0.6",
-    "react": "^16.3.0",
-    "react-dom": "^16.3.0",
-    "react-map-gl": "^3.3.0"
-  },
-  "devDependencies": {
-    "buble": "^0.19.3",
-    "buble-loader": "^0.5.0",
-    "html-webpack-plugin": "^3.0.7",
-    "webpack": "^4.20.2",
-    "webpack-cli": "^3.1.2",
-    "webpack-dev-server": "^3.1.1"
-  }
-}
