--- test/interaction/template.html
@@ -1,19 +1,0 @@
-<!doctype html>
-<html lang="en">
-<head>
-  <meta charset="utf-8">
-  <title>deck.gl Example</title>
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-  <style>
-    body {margin: 0; font-family: sans-serif; width: 100vw; height: 100vh; overflow: hidden;}
-    .tooltip {pointer-events: none; position: absolute; z-index: 9; font-size: 12px; padding: 8px; background: #000; color: #fff; min-width: 160px; max-height: 240px; overflow-y: hidden;}
-    .tooltip.interactive {pointer-events: all; background: #fcfcfc; color: #444; overflow-y: auto;}
-  </style>
-</head>
-<body>
-  <div id="app"></div>
-</body>
-<script type="text/javascript">
-  // App.renderToDOM(document.getElementById('app'));
-</script>
-</html>
