--- test/interaction/webpack.config.js
@@ -1,163 +1,0 @@
-// This file contains webpack configuration settings that allow
-// examples to be built against the deck.gl source code in this repo instead
-// of building against their installed version of deck.gl.
-//
-// This enables using the examples to debug the main deck.gl library source
-// without publishing or npm linking, with conveniences such hot reloading etc.
-
-// avoid destructuring for older Node version support
-const resolve = require('path').resolve;
-
-const LIB_DIR = resolve(__dirname, '../..');
-const SRC_DIR = resolve(LIB_DIR, './modules');
-
-const ALIASES = require('../../aliases')('src');
-
-const webpack = require('webpack');
-const HtmlWebpackPlugin = require('html-webpack-plugin');
-
-// Support for hot reloading changes to the deck.gl library:
-function makeLocalDevConfig(EXAMPLE_DIR = LIB_DIR) {
-  return {
-    // TODO - Uncomment when all examples use webpack 4 for faster bundling
-    // mode: 'development',
-
-    // suppress warnings about bundle size
-    devServer: {
-      stats: {
-        warnings: false
-      }
-    },
-
-    devtool: 'source-map',
-
-    resolve: {
-      // mainFields: ['esnext', 'module', 'main'],
-
-      alias: Object.assign({}, ALIASES, {
-        // Use luma.gl specified by root package.json
-        'luma.gl': resolve(LIB_DIR, './node_modules/luma.gl'),
-        // Important: ensure shared dependencies come from the main node_modules dir
-        // Versions will be controlled by the deck.gl top level package.json
-        'math.gl': resolve(LIB_DIR, './node_modules/math.gl'),
-        'viewport-mercator-project': resolve(LIB_DIR, './node_modules/viewport-mercator-project'),
-        seer: resolve(LIB_DIR, './node_modules/seer'),
-        react: resolve(LIB_DIR, './node_modules/react')
-      })
-    },
-    module: {
-      rules: [
-        {
-          // Unfortunately, webpack doesn't import library sourcemaps on its own...
-          test: /\.js$/,
-          use: ['source-map-loader'],
-          enforce: 'pre'
-        }
-      ]
-    }
-  };
-}
-
-const BUBLE_CONFIG = {
-  module: {
-    rules: [
-      {
-        // Compile source using buble
-        test: /\.js$/,
-        loader: 'buble-loader',
-        include: [SRC_DIR],
-        options: {
-          objectAssign: 'Object.assign',
-          transforms: {
-            dangerousForOf: true,
-            modules: false
-          }
-        }
-      }
-    ]
-  }
-};
-
-const CONFIG = {
-  mode: 'development',
-
-  entry: {
-    app: resolve('./app.js')
-  },
-
-  output: {
-    library: 'App'
-  },
-
-  module: {
-    rules: [
-      {
-        // Compile ES2015 using buble
-        test: /\.js$/,
-        loader: 'buble-loader',
-        include: [resolve('.')],
-        exclude: [/node_modules/],
-        options: {
-          objectAssign: 'Object.assign'
-        }
-      }
-    ]
-  },
-
-  resolve: {
-    alias: {
-      // From mapbox-gl-js README. Required for non-browserify bundlers (e.g. webpack):
-      'mapbox-gl$': resolve('./node_modules/mapbox-gl/dist/mapbox-gl.js')
-    }
-  },
-
-  // Optional: Enables reading mapbox token from environment variable
-  plugins: [
-    // new HtmlWebpackPlugin({title: 'deck.gl example'})
-    new HtmlWebpackPlugin({template: './template.html'}),
-    new webpack.EnvironmentPlugin(['MapboxAccessToken'])
-  ]
-};
-
-function addLocalDevSettings(config, exampleDir) {
-  const LOCAL_DEV_CONFIG = makeLocalDevConfig(exampleDir);
-  config = Object.assign({}, LOCAL_DEV_CONFIG, config);
-  config.resolve = Object.assign({}, LOCAL_DEV_CONFIG.resolve, config.resolve || {});
-  config.resolve.alias = config.resolve.alias || {};
-  Object.assign(config.resolve.alias, LOCAL_DEV_CONFIG.resolve.alias);
-
-  config.module = config.module || {};
-  Object.assign(config.module, {
-    rules: (config.module.rules || []).concat(LOCAL_DEV_CONFIG.module.rules)
-  });
-  return config;
-}
-
-function addBubleSettings(config) {
-  config.module = config.module || {};
-  Object.assign(config.module, {
-    rules: (config.module.rules || []).concat(BUBLE_CONFIG.module.rules)
-  });
-  return config;
-}
-
-module.exports = env => {
-  // npm run start-local now transpiles the lib
-  let config = Object.assign({}, CONFIG);
-  if (env && env.local) {
-    config = addLocalDevSettings(config);
-    config = addBubleSettings(config);
-  }
-
-  // npm run start-es6 does not transpile the lib
-  if (env && env.es6) {
-    config = addLocalDevSettings(config);
-  }
-
-  if (env && env.production) {
-    config.mode = 'production';
-  }
-
-  // console.warn(JSON.stringify(config, null, 2)); // uncomment to debug config
-  return config;
-};
